Add files via upload
Added the possibility to pass parameters through command line. No reading of a file is needed. Using this options will also not write into a file but print out the values that would normally go into int.dat. This is used for spinteractive.
This mode can be activated by using -x as the first parameter and then supplying one long string. See example in documentation.